requirements.txt
typer
requests
rich
openai
python-dotenv
